refactor: コードの整理と関数のリファクタリングを実施し、全問題取得機能を追加
