+++ .gitignore
@@ -15,6 +15,7 @@ node_modules
 # production
 build
 dist
+lib
 
 # misc
 .DS_Store
